Remove call to `authorizeUser` from constructor
The original thought of having the call to `authorizeUser` in the constructor was that the GitJs object would attempt to authenticate the user if the mode was set to 'write'. This really doesn't work since the call to `authorizeUser` is asynchronous and would cause a race condition. While the `inheriting` variable is no longer used, I'm leaving it in because it will be used in the future.

This closes #32.