Set up example for how to use userAuthenticate method
I set up the basic layout for the example page on how to use the
authenticateUser method (see #46). Right now it just pulls in the
user's profile pic as well as there name, email and website. There's
still more that needs to be done in terms of polish, but I feel it's a
good start and a good foundation to build on.