Add logic to handle when permissions are changed and when token is already present
Added logic that checks to see if the user has changed the level of
access the application should have to their account. If it has changed,
it requests a new token. I also added logic that checks to see if a
token is already present. If it is present, it simply logs the user in
instead of requesting a new token.